Reintroduce old implementation of product datasource and put new implementation into own datasource for compability
